Add new events components to existing events folder and rename some
