+++ app/web/features/communities/events/CommunityEventsList.stories.tsx
@@ -3,20 +3,20 @@ import { mockedService } from "stories/serviceMocks";
 import community from "test/fixtures/community.json";
 import events from "test/fixtures/events.json";
 
-import EventsList from "./EventsList";
+import CommunityEventsList from "./CommunityEventsList";
 
 export default {
-  component: EventsList,
-  title: "Communities/Events/EventsList",
+  component: CommunityEventsList,
+  title: "Communities/Events/CommunityEventsList",
 } as Meta;
 
-interface EventsListArgs {
+interface CommunityEventsListArgs {
   pagesOfEvent?: "none" | "one-page" | "multiple-pages";
   shouldListAttendeesSucceed?: boolean;
   shouldListEventsSucceed?: boolean;
 }
 
-const Template: Story<EventsListArgs> = ({
+const Template: Story<CommunityEventsListArgs> = ({
   pagesOfEvent = "one-page",
   shouldListAttendeesSucceed = true,
   shouldListEventsSucceed = true,
@@ -26,7 +26,7 @@ const Template: Story<EventsListArgs> = ({
     shouldListAttendeesSucceed,
     shouldListEventsSucceed,
   });
-  return <EventsList community={community} />;
+  return <CommunityEventsList community={community} />;
 };
 
 export const OnePageOfEvents = Template.bind({});
@@ -49,7 +49,7 @@ ErrorLoadingEvents.args = {
 function setMocks({
   pagesOfEvent,
   shouldListEventsSucceed,
-}: Required<EventsListArgs>) {
+}: Required<CommunityEventsListArgs>) {
   mockedService.events.listCommunityEvents = async () => {
     return shouldListEventsSucceed
       ? {
