+++ app/web/features/communities/events/LongEventCard.tsx
@@ -1,145 +1,164 @@
 import {
   Card,
+  CardContent,
+  CardMedia,
   Theme,
+  Tooltip,
   Typography,
-  useMediaQuery,
-  useTheme,
 } from "@material-ui/core";
 import { eventImagePlaceholderUrl } from "appConstants";
-import { AttendeesIcon, CalendarIcon } from "components/Icons";
+import Pill from "components/Pill";
 import { useTranslation } from "i18n";
 import { COMMUNITIES } from "i18n/namespaces";
 import Link from "next/link";
 import { Event } from "proto/events_pb";
-import { useMemo } from "react";
 import { routeToEvent } from "routes";
+import { theme } from "theme";
 import { timestamp2Date } from "utils/date";
 import dayjs from "utils/dayjs";
 import makeStyles from "utils/makeStyles";
 
-import getContentSummary from "../getContentSummary";
-
-const useStyles = makeStyles<Theme, { eventImageSrc: string }>((theme) => ({
-  overviewRoot: {
-    "&:hover": {
-      backgroundColor: theme.palette.grey[50],
+const useStyles = makeStyles((theme: Theme) => ({
+  root: {
+    margin: 0,
+    "&:not(:first-child)": {
+      margin: theme.spacing(2, 0),
     },
-    padding: theme.spacing(1),
-    [theme.breakpoints.up("sm")]: {
-      padding: theme.spacing(2),
-    },
+    border: `1px solid ${theme.palette.grey[300]}`,
+    borderRadius: theme.spacing(1),
   },
-  overviewContent: {
-    display: "grid",
-    gap: theme.spacing(1),
-    gridTemplateAreas: `
-      "titles image"
-      "content content"
-    `,
-    gridTemplateColumns: "1fr 1fr",
-    [theme.breakpoints.up("md")]: {
-      gridTemplateColumns: "2fr 1fr",
-    },
-  },
-  eventInfoContainer: {
-    gridArea: "titles",
-  },
-  eventTimeContainer: {
-    alignItems: "center",
+  attendees: {
     display: "flex",
-    marginBlockStart: theme.spacing(1),
+    alignItems: "flex-end",
+    justifyContent: "flex-end",
+    minWidth: theme.spacing(10),
   },
-  attendeesCountContainer: {
-    alignItems: "center",
+  card: {
     display: "flex",
-  },
-  icon: {
-    display: "block",
-    fontSize: "1.25rem",
-    lineHeight: 1.5,
-    marginInlineEnd: theme.spacing(0.5),
-  },
-  image: {
-    objectFit: ({ eventImageSrc }) =>
-      eventImageSrc === eventImagePlaceholderUrl ? "contain" : "cover",
-    height: 80,
-    [theme.breakpoints.up("md")]: {
-      height: 120,
-    },
     width: "100%",
-    gridArea: "image",
+    height: theme.spacing(20),
+    [theme.breakpoints.down("xs")]: {
+      height: "auto",
+    },
   },
-  onlineOrOfflineInfo: {
-    fontWeight: "bold",
-    color: theme.palette.grey[600],
+  cardMedia: {
+    height: "100%",
+    width: "25%",
+    objectFit: "fill",
   },
-  content: {
-    gridArea: "content",
+  cardContent: {
+    width: "75%",
+    display: "flex",
+    flexDirection: "column",
+    justifyContent: "space-between",
+  },
+  eventInfo: {
+    display: "flex",
+    justifyContent: "flex-end",
+    flexDirection: "column",
+  },
+  row: {
+    display: "flex",
+    justifyContent: "space-between",
+  },
+  tags: {
+    minWidth: theme.spacing(15),
+    [theme.breakpoints.down("xs")]: {
+      minWidth: theme.spacing(10),
+    },
+  },
+  title: {
+    display: "-webkit-box",
+    lineClamp: 2,
+    boxOrient: "vertical",
+    overflow: "hidden",
+    textOverflow: "ellipsis",
+    maxHeight: "3em" /* Approximate height for 2 lines of text */,
+    lineHeight: "1.5em",
+    paddingRight: theme.spacing(2),
   },
 }));
 
-export interface LongEventCardProps {
+const LongEventCard = ({
+  event,
+  userId,
+}: {
   event: Event.AsObject;
-}
-
-export default function LongEventCard({ event }: LongEventCardProps) {
-  const { t } = useTranslation([COMMUNITIES]);
+  userId: number | null | undefined;
+}) => {
   const classes = useStyles({
     eventImageSrc: event.photoUrl || eventImagePlaceholderUrl,
   });
-  const theme = useTheme();
-  const isBelowLg = useMediaQuery(theme.breakpoints.down("md"));
+  const { t } = useTranslation([COMMUNITIES]);
 
-  const truncatedContent = useMemo(
-    () =>
-      getContentSummary({
-        originalContent: event.content,
-        maxLength: isBelowLg ? 120 : 300,
-      }),
-    [event.content, isBelowLg]
-  );
-  const startTime = dayjs(timestamp2Date(event.startTime!));
+  const startTime = dayjs(timestamp2Date(event.startTime!)).format("llll");
+
+  const renderTags = () => {
+    const isCreatedByMe = event.creatorUserId === userId;
+    const isOnline = event.onlineInformation?.link !== undefined;
+    const isCancelled = event.isCancelled;
+
+    return (
+      <div className={classes.tags}>
+        {isCreatedByMe && (
+          <Pill variant="rounded">{t("communities:created_by_me")}</Pill>
+        )}
+        {isOnline && <Pill variant="rounded">{t("communities:online")}</Pill>}
+        {isCancelled && (
+          <Pill
+            backgroundColor={theme.palette.error.main}
+            color={theme.palette.common.white}
+            variant="rounded"
+          >
+            {t("communities:cancelled")}
+          </Pill>
+        )}
+      </div>
+    );
+  };
 
   return (
-    <Card className={classes.overviewRoot}>
+    <Card className={classes.root} data-testid="event-item">
       <Link href={routeToEvent(event.eventId, event.slug)}>
-        <a className={classes.overviewContent}>
-          <div className={classes.eventInfoContainer}>
-            <Typography variant="h2">{event.title}</Typography>
-            <Typography
-              className={classes.onlineOrOfflineInfo}
-              color="primary"
-              variant="body1"
-            >
-              {event.offlineInformation
-                ? event.offlineInformation.address
-                : t("communities:online")}
-            </Typography>
-            <div className={classes.eventTimeContainer}>
-              <CalendarIcon className={classes.icon} />
-              <Typography variant="body1">
-                {startTime.format("ll LT")}
-              </Typography>
+        <a className={classes.card}>
+          <CardMedia
+            className={classes.cardMedia}
+            component="img"
+            image={event.photoUrl || eventImagePlaceholderUrl}
+          />
+          <CardContent className={classes.cardContent}>
+            <div className={classes.row}>
+              <Tooltip title={event.title}>
+                <Typography variant="h3" className={classes.title}>
+                  {event.title}
+                </Typography>
+              </Tooltip>
+              {renderTags()}
             </div>
-            <div className={classes.attendeesCountContainer}>
-              <AttendeesIcon className={classes.icon} />
-              <Typography variant="body1">
+            <div className={classes.row}>
+              <div className={classes.eventInfo}>
+                <Typography variant="body2">
+                  {event.offlineInformation
+                    ? event.offlineInformation.address
+                    : t("communities:virtual_event_location_placeholder")}
+                </Typography>
+                <Typography variant="body2">{startTime}</Typography>
+              </div>
+              <Typography
+                className={classes.attendees}
+                variant="body2"
+                color="textSecondary"
+              >
                 {t("communities:attendees_count", {
                   count: event.goingCount + event.maybeCount,
                 })}
               </Typography>
             </div>
-          </div>
-          <Typography className={classes.content} variant="body1">
-            {truncatedContent}
-          </Typography>
-          <img
-            alt=""
-            className={classes.image}
-            src={event.photoUrl || eventImagePlaceholderUrl}
-          />
+          </CardContent>
         </a>
       </Link>
     </Card>
   );
-}
+};
+
+export default LongEventCard;
